Default Home to PublicHome to reduce wait
We don't expect users to be logged in the first time they visit the site.
Instead of waiting to see if they're logged in let's just show
PublicHome and redirect to /dashboard once we know they're logged in.

Locally this improved performance from 74 to 83 with the major
improvement being the reduction of "Largest Contentful Paint" from 5.1s to 3.6s.